requirements-simplified.txt
# Core AI Framework
crewai==0.11.2
langchain>=0.1.0
langchain-openai>=0.0.5
openai>=1.12.0

# Vector Database
pinecone-client>=3.0.0

# WebSocket Support
websockets>=12.0
python-socketio>=5.11.0

# Web Search
tavily-python>=0.3.0

# API Framework
fastapi>=0.108.0
uvicorn[standard]>=0.25.0
pydantic>=2.5.0
python-multipart>=0.0.6

# Database and Storage
redis>=5.0.1
aioredis>=2.0.1

# Data Processing
pandas>=2.1.4
numpy>=1.26.2
tiktoken>=0.5.2

# Web Scraping
beautifulsoup4>=4.12.2
httpx>=0.26.0

# Utilities
python-dotenv>=1.0.0
aiofiles>=23.2.1

# Monitoring and Logging
loguru>=0.7.2